feat(builder): use temp local registry for testing multi-platform builds.
